fixes, getUTXOs strategy, unit tests
